Updated remaining PATCH and DELETE controllers for events on backend

+++ server/services/events_service.py
@@ -22,9 +22,9 @@ class EventService:
     return Event.query.join(Job).filter(Job.user_id==user_id)
   
   @staticmethod
-  def get_event_for_job(event_id, job_id):
-    event = Event.query.filter_by(id=event_id, job_id=job_id).first()
-    return event
+  def get_events_for_job(event_id, job_id):
+    events = Event.query.filter_by(id=event_id, job_id=job_id).all()
+    return events
   
   @staticmethod
   def create_event(job_id, data):
@@ -52,25 +52,38 @@ class EventService:
       return None, "invalid_data"
     
   @staticmethod
-  def update_event(event, data):
+  def update_event(event_id, data):
+    event = Event.query.filter_by(id=event_id).first()
+    
+    if not event:
+      return None, "event_not_found"
     for key, value in data.items():
+      if key == "scheduled_time":
+        try:
+          value = datetime.fromisoformat(value)  # convert string → datetime
+        except ValueError:
+          return None, "invalid_data"
       setattr(event, key, value)
     try:
       db.session.commit()
-      return event
+      return event, None
     except IntegrityError:
       db.session.rollback()
-      return None
+      return None, "invalid_data"
     
   @staticmethod
-  def delete_event(event):
+  def delete_event(event_id):
+    event = Event.query.filter_by(id=event_id).first()
+    if not event:
+      return False, "event_not_found"
+  
     try:
       db.session.delete(event)
       db.session.commit()
-      return True
+      return True, None
     except IntegrityError:
       db.session.rollback()
-      return False
+      return False, "invalid_data"
     
 # 401 Unauthorized handled by jwt_required
 # 404 Not Found
